ProjectImplemmentation

+++ src/main/java/com/ii/testautomation/service/impl/TestGroupingServiceImpl.java
@@ -16,6 +16,7 @@ import org.springframework.beans.BeanUtils;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.data.domain.Page;
+import org.springframework.data.domain.PageImpl;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 import org.springframework.web.multipart.MultipartFile;
@@ -198,7 +199,9 @@ public class TestGroupingServiceImpl implements TestGroupingService {
         File newGroupFolder = new File(newGroupFolderPath);
         List<String> excelPaths = testGrouping.getExcelFilePath();
         List<String> newExcelPathList = new ArrayList<>();
-        existingGroupFolder.renameTo(newGroupFolder);
+        if (existingGroupFolder.exists()) {
+            existingGroupFolder.renameTo(newGroupFolder);
+        }
         testGrouping.setGroupPath(newGroupFolderPath);
         if (excelPaths != null && !excelPaths.isEmpty()) {
             for (String excelPath : excelPaths
@@ -340,41 +343,90 @@ public class TestGroupingServiceImpl implements TestGroupingService {
     @Override
     public List<TestGroupingResponse> getAllTestGroupingByProjectId(Pageable pageable, PaginatedContentResponse.Pagination pagination, Long projectId) {
         List<TestGroupingResponse> testGroupingResponseList = new ArrayList<>();
-        Page<TestGrouping> testGroupingPage = testGroupingRepository.findDistinctTestGroupingByTestCases_SubModule_MainModule_Modules_Project_Id(pageable, projectId);
-        List<String> testCaseNames = new ArrayList<>();
-        List<String> testScenariosNames = new ArrayList<>();
-        List<Long> testScenariosIds = new ArrayList<>();
-        List<Long> testCaseIds = new ArrayList<>();
-        pagination.setTotalRecords(testGroupingPage.getTotalElements());
-        pagination.setPageSize(testGroupingPage.getTotalPages());
+        BooleanBuilder booleanBuilder = new BooleanBuilder();
+        BooleanBuilder testScenariosbooleanBuilder = new BooleanBuilder();
+        QTestGrouping qTestGrouping = QTestGrouping.testGrouping;
+
+        if (qTestGrouping.testCases != null &&
+                qTestGrouping.testCases.any().subModule != null &&
+                qTestGrouping.testCases.any().subModule.mainModule != null &&
+                qTestGrouping.testCases.any().subModule.mainModule.modules != null &&
+                qTestGrouping.testCases.any().subModule.mainModule.modules.name != null) {
+            booleanBuilder.and(qTestGrouping.testCases.any().subModule.mainModule.modules.project.id.eq(projectId));
+
+        }
+        Page<TestGrouping> testGroupingPageByTestCase = testGroupingRepository.findAll(booleanBuilder, pageable);
+
+        if (qTestGrouping.testScenarios != null &&
+                qTestGrouping.testScenarios.any().testCases != null &&
+                qTestGrouping.testScenarios.any().testCases.any().subModule != null &&
+                qTestGrouping.testScenarios.any().testCases.any().subModule.mainModule != null &&
+                qTestGrouping.testScenarios.any().testCases.any().subModule.mainModule.modules != null &&
+                qTestGrouping.testScenarios.any().testCases.any().subModule.mainModule.modules.name != null) {
+            testScenariosbooleanBuilder.and(qTestGrouping.testCases.any().subModule.mainModule.modules.project.id.eq(projectId));
+
+        }
+        Page<TestGrouping> testGroupingPageByTestScenarios = testGroupingRepository.findAll(testScenariosbooleanBuilder, pageable);
+        Page<TestGrouping> testGroupingPage = combineAndRemoveDuplicates(testGroupingPageByTestCase, testGroupingPageByTestScenarios);
 
         for (TestGrouping testGrouping : testGroupingPage) {
             TestGroupingResponse testGroupingResponse = new TestGroupingResponse();
-            testGroupingResponse.setTestTypeName(testGrouping.getTestType().getName());
-            testGroupingResponse.setName(testGrouping.getName());
-            testGroupingResponse.setId(testGrouping.getId());
-            for (TestCases testCases : testGrouping.getTestCases()) {
-                testCaseNames.add(testCases.getName());
-                testCaseIds.add(testCases.getId());
+            if (testGrouping.getTestType() != null) {
+                testGroupingResponse.setTestTypeName(testGrouping.getTestType().getName());
+                testGroupingResponse.setTestTypeId(testGrouping.getTestType().getId());
             }
-            for (TestScenarios testScenarios : testGrouping.getTestScenarios()) {
-                testScenariosNames.add(testScenarios.getName());
-                testScenariosIds.add(testScenarios.getId());
+            if (testGrouping.getName() != null) {
+                testGroupingResponse.setName(testGrouping.getName());
             }
-            testGroupingResponse.setTestTypeId(testGrouping.getTestType().getId());
+            if (testGrouping.getId() != null) {
+                testGroupingResponse.setId(testGrouping.getId());
+            }
+            List<String> testCaseNames = new ArrayList<>();
+            List<Long> testCaseIds = new ArrayList<>();
+            if (testGrouping.getTestCases() != null && !testGrouping.getTestCases().isEmpty()) {
+                for (TestCases testCases : testGrouping.getTestCases()) {
+                    if (testCases.getName() != null) {
+                        testCaseNames.add(testCases.getName());
+                    }
+                    if (testCases.getId() != null) {
+                        testCaseIds.add(testCases.getId());
+                    }
+                }
+            }
+            List<String> testScenariosNames = new ArrayList<>();
+            List<Long> testScenariosIds = new ArrayList<>();
+            if (testGrouping.getTestScenarios() != null && !testGrouping.getTestScenarios().isEmpty()) {
+                for (TestScenarios testScenarios : testGrouping.getTestScenarios()) {
+                    if (testScenarios.getName() != null) {
+                        testScenariosNames.add(testScenarios.getName());
+                    }
+                    if (testScenarios.getId() != null) {
+                        testScenariosIds.add(testScenarios.getId());
+                    }
+                }
+            }
             List<String> sortedTestCaseNames = testCaseNames.stream().distinct().collect(Collectors.toList());
             List<String> sortedTestScenarioNames = testScenariosNames.stream().distinct().collect(Collectors.toList());
             List<Long> sortedTestScenariosIds = testScenariosIds.stream().distinct().collect(Collectors.toList());
             List<Long> sortedTestCasesIds = testCaseIds.stream().distinct().collect(Collectors.toList());
+
             testGroupingResponse.setTestCaseIds(sortedTestCasesIds);
-            testGroupingResponse.setTestScenarioName(sortedTestScenarioNames);
             testGroupingResponse.setTestScenarioIds(sortedTestScenariosIds);
             testGroupingResponse.setTestCaseName(sortedTestCaseNames);
+            testGroupingResponse.setTestScenarioName(sortedTestScenarioNames);
             testGroupingResponseList.add(testGroupingResponse);
         }
 
-
         return testGroupingResponseList;
+    }
+
+    private Page<TestGrouping> combineAndRemoveDuplicates(Page<TestGrouping> page1, Page<TestGrouping> page2) {
+        Set<TestGrouping> uniqueTestGroupings = new HashSet<>(page1.getContent());
+        uniqueTestGroupings.addAll(page2.getContent());
+
+        List<TestGrouping> combinedContent = new ArrayList<>(uniqueTestGroupings);
+
+        return new PageImpl<>(combinedContent, page1.getPageable(), combinedContent.size());
     }
 
     @Override
@@ -455,41 +507,16 @@ public class TestGroupingServiceImpl implements TestGroupingService {
         }
     }
 
-    //    @Override
-//    public void updateTestGroupingExecutionStatus(Long testGroupingId, Long projectId, List<Long> testScenarioIds, List<Long> testCaseIds) {
-//        TestGrouping testGrouping = testGroupingRepository.findById(testGroupingId).orElse(null);
-//        testGrouping.setExecutionStatus(true);
-//        testGroupingRepository.save(testGrouping);
-//        if (testScenarioIds != null && !testScenarioIds.isEmpty()) {
-//            for (Long testScenarioId : testScenarioIds
-//            ) {
-//                TestScenarios testScenarios = testScenarioRepository.findById(testScenarioId).get();
-//                testScenarios.setExecutionStatus(true);
-//                testScenarioRepository.save(testScenarios);
-//            }
-//        }
-//        if (testCaseIds != null && !testCaseIds.isEmpty()) {
-//            for (Long testCaseId : testCaseIds
-//            ) {
-//                TestCases testCases = testCasesRepository.findById(testCaseId).get();
-//                testCases.setExecutionStatus(true);
-//                testCasesRepository.save(testCases);
-//            }
-//        }
-//        String savedFilePath = projectRepository.findById(projectId).get().getJarFilePath();
-//        File jarFile = new File(savedFilePath);
-//        String jarFileName = jarFile.getName();
-//        String jarDirectory = jarFile.getParent();
-//        try {
-//            ProcessBuilder runProcessBuilder = new ProcessBuilder("java", "-jar", jarFileName);
-//            runProcessBuilder.directory(new File(jarDirectory));
-//            runProcessBuilder.redirectErrorStream(true);
-//            Process runProcess = runProcessBuilder.start();
-//            runProcess.waitFor();
-//        } catch (IOException | InterruptedException e) {
-//            e.printStackTrace();
-//        }
-//    }
+    @Override
+    public boolean folderExists(Long groupId) {
+        String groupPath = testGroupingRepository.findById(groupId).get().getGroupPath();
+        if (groupPath != null) {
+            File file = new File(groupPath);
+            if (file.exists()) return true;
+        }
+        return false;
+    }
+
     @Override
     public boolean existsByTestGroupingNameByProjectId(String name, Long projectId) {
         return testGroupingRepository.existsByNameIgnoreCaseAndTestCases_SubModule_MainModule_Modules_Project_Id(name, projectId);
@@ -569,7 +596,8 @@ public class TestGroupingServiceImpl implements TestGroupingService {
     }
 
     @Override
-    public List<TestGroupingResponse> multiSearchTestGrouping(Pageable pageable, PaginatedContentResponse.Pagination pagination, TestGroupingSearch testGroupingSearch) {
+    public List<TestGroupingResponse> multiSearchTestGrouping(Pageable
+                                                                      pageable, PaginatedContentResponse.Pagination pagination, TestGroupingSearch testGroupingSearch) {
         BooleanBuilder booleanBuilder = new BooleanBuilder();
         if (Utils.isNotNullAndEmpty(testGroupingSearch.getName())) {
             booleanBuilder.and(QTestGrouping.testGrouping.name.containsIgnoreCase(testGroupingSearch.getName()));
